Move test contracts from `modules/` to `contracts/`
